% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidjsonr-package.R
\docType{package}
\name{rapidjsonr-package}
\alias{rapidjsonr}
\alias{rapidjsonr-package}
\title{This package provides rapidjson headers for parsing JSON}
\description{
This package provides rapidjson headers for parsing JSON
}
\author{
\strong{Maintainer}: David Cooley \email{dcooley@symbolix.com.au}

}
