% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidjsonr-package.R
\docType{package}
\name{rapidjsonr-package}
\alias{rapidjsonr}
\alias{rapidjsonr-package}
\title{rapidjsonr: Rapidjson C++ Header Files}
\description{
Rapidjson provides a c++ header-only library for parsing GeoJSON.
This package exposes those header files for use in R libraries.
}
\author{
\strong{Maintainer}: David Cooley \email{dcooley@symbolix.com.au}

}
