% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidjsonr-package.R
\docType{package}
\name{rapidjsonr-package}
\alias{rapidjsonr}
\alias{rapidjsonr-package}
\title{Proivdes JSON parsing capability through the 'Rapidjson' 'C++' header-only library.}
\description{
Proivdes JSON parsing capability through the 'Rapidjson' 'C++' header-only library.
}
\examples{
\dontrun{

# Include a dependency to rapidjsonr inside a cpp file
// [[Rcpp::depends(rapidjsonr)]]
include <Rcpp.h>
}
}
\author{
\strong{Maintainer}: David Cooley \email{dcooley@symbolix.com.au}

Other contributors:
\itemize{
  \item Milo Yip (Author of c++ rapidjson library, provided through THL A29 Limited, a Tencent company) [contributor]
  \item Alexander Chemeris (Author of c++ msinttypes library) [contributor]
}

}
